fix:修复美团团购券 实际支付金额错误的问题

+++ mazongjian-module-member/mazongjian-module-member-biz/src/main/java/com/yanzu/module/member/controller/app/chart/AppChartController.java
@@ -98,7 +98,7 @@ public class AppChartController {
     @PostMapping("/getRoomUseHourStatistics")
     @Operation(summary = "获取房间使用时长统计")
     @PreAuthenticated
-    public CommonResult<List<KeyValue<String, Double>>> getRoomUseHourStatistics(@RequestBody AppChartDataReqVO reqVO) {
+    public CommonResult<List<KeyValue<String, BigDecimal>>> getRoomUseHourStatistics(@RequestBody AppChartDataReqVO reqVO) {
         return success(appMangerService.getRoomUseHourStatistics(reqVO));
     }
 }
