+++ mazongjian-module-member/mazongjian-module-member-biz/src/main/java/com/yanzu/module/member/dal/mysql/orderinfo/OrderInfoMapper.java
@@ -49,7 +49,7 @@ public interface OrderInfoMapper extends BaseMapperX<OrderInfoDO> {
 
     List<KeyValue<String, Long>> getRoomUseStatistics(AppChartDataReqVO reqVO);
 
-    List<KeyValue<String, Double>> getRoomUseHourStatistics(AppChartDataReqVO reqVO);
+    List<KeyValue<String, BigDecimal>> getRoomUseHourStatistics(AppChartDataReqVO reqVO);
 
     List<OrderInfoDO> getByStatus(Integer status);
