+++ mazongjian-module-member/mazongjian-module-member-biz/src/main/java/com/yanzu/module/member/service/manager/AppMangerService.java
@@ -57,7 +57,7 @@ public interface AppMangerService {
 
     List<KeyValue<String, String>> getRoomUseStatistics(AppChartDataReqVO reqVO);
 
-    List<KeyValue<String, Double>> getRoomUseHourStatistics(AppChartDataReqVO reqVO);
+    List<KeyValue<String, BigDecimal>> getRoomUseHourStatistics(AppChartDataReqVO reqVO);
 
     void giftCoupon(AppGiftCouponReqVO reqVO);
